favorite/permissions.py
from rest_framework import permissions


class IsAuthenticatedOrReadOnly(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):

        if request.method in permissions.SAFE_METHODS and request.user == obj.author:
            return True
        
        return request.user == obj.author